% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EasyCalibrator-package.R
\docType{package}
\name{EasyCalibrator-package}
\alias{EasyCalibrator}
\alias{EasyCalibrator-package}
\title{EasyCalibrator: Calibrate TB/HIV Model}
\description{
More about what it does (maybe more than one line)
    Use four spaces when indenting paragraphs within the Description.
}
\keyword{internal}
